int to float

+++ SumOfTwoNumbers.py
@@ -1,3 +1,3 @@
-num1 = int(input("Enter 1st Number : "))
-num2 = int(input("Enter 2nd Number : "))
+num1 = float(input("Enter 1st Number : "))
+num2 = float(input("Enter 2nd Number : "))
 print(f"Sum of your number is : {num1+num2}")
